Draft `DeltaUnpack16`
Converting the sse instructions here into intrinsics to simulate the original implementation.
Seems to work well enough, though when rows are the same(such as when the tile is filled completely with the color red or with a vertical strike pattern) the subsequent rows seem to just be zeroes.
There is some optimization happening somewhere that allows duplicate rows of tile data to not have to be stored again.