Quit at the first found supported thumbnail
At this point, the older thumbnail format is proving to be quite rare, while the jpeg-stream based format is so much more common in my small "survey". So I think it's worth doing a public release with just the jssf format for now and supporting the older png-filter based one later.